+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Removed mapping of `id` to `_id` with `order` option
 - Removed `wordnet` option
 - Removed `elasticsearch` dependency
+- Dropped support for Elasticsearch 6
 - Dropped support for Ruby < 2.6 and Active Record < 5.2
 - Dropped support for NoBrainer and Cequel
 - Dropped support for `faraday_middleware-aws-signers-v4` (use `faraday_middleware-aws-sigv4` instead)